requirements.txt
torch==2.11.0
torchvision==0.26.0
numpy<2
pandas==2.2.3
scikit-learn==1.8.0
matplotlib==3.8.3
Pillow==10.3.0
scipy==1.15.2
scikit-image==0.25.2
opencv-python-headless>=4.10.0
huggingface_hub>=1.0.0
